+++ Rubeus/Commands/Asktgt.cs
@@ -40,38 +40,59 @@ namespace Rubeus.Commands
             {
                 dc = arguments["/dc"];
             }
+
             if (arguments.ContainsKey("/password"))
             {
                 password = arguments["/password"];
-                if (arguments.ContainsKey("/enctype") && arguments["/enctype"].ToUpper().Equals("AES256"))
-                {
-                    encType = Interop.KERB_ETYPE.aes256_cts_hmac_sha1;
 
-                    // compute AES key from pwd
-                    byte[] password_bytes = System.Text.Encoding.UTF8.GetBytes(password);
-                    byte[] salt = System.Text.Encoding.UTF8.GetBytes(domain.ToUpper() + user);
+                string salt = String.Format("{0}{1}", domain.ToUpper(), user.ToLower());
+                encType = Interop.KERB_ETYPE.rc4_hmac; //default is non /enctype is specified
 
-                    byte[] aes256_key = Crypto.ComputeAES256KerberosKey(password_bytes, salt);
-                    hash = System.BitConverter.ToString(aes256_key).Replace("-", "");
-                }
-                else // default is RC4
+                if (arguments.ContainsKey("/enctype"))
                 {
-                    // compute NTLM from pwd
-                    encType = Interop.KERB_ETYPE.rc4_hmac;
-                    byte[] ntlm = Crypto.ComputeRC4KerberosKey(password); // a.k.a NTLM
-                    hash = System.BitConverter.ToString(ntlm).Replace("-", "");
+                    string encTypeString = arguments["/enctype"].ToUpper();
+
+                    if (encTypeString.Equals("RC4"))
+                    {
+                        encType = Interop.KERB_ETYPE.rc4_hmac;
+                    }
+                    else if (encTypeString.Equals("AES128"))
+                    {
+                        encType = Interop.KERB_ETYPE.aes128_cts_hmac_sha1;
+                    }
+                    else if (encTypeString.Equals("AES256") || encTypeString.Equals("AES"))
+                    {
+                        encType = Interop.KERB_ETYPE.aes256_cts_hmac_sha1;
+                    }
+                    else if (encTypeString.Equals("DES"))
+                    {
+                        encType = Interop.KERB_ETYPE.des_cbc_md5;
+                    }
                 }
+                hash = Crypto.KerberosPasswordHash(encType, password, salt);
             }
-            if (arguments.ContainsKey("/rc4"))
+
+            else if (arguments.ContainsKey("/des"))
+            {
+                hash = arguments["/des"];
+                encType = Interop.KERB_ETYPE.des_cbc_md5;
+            }
+            else if (arguments.ContainsKey("/rc4"))
             {
                 hash = arguments["/rc4"];
                 encType = Interop.KERB_ETYPE.rc4_hmac;
             }
-            if (arguments.ContainsKey("/aes256"))
+            else if (arguments.ContainsKey("/aes128"))
+            {
+                hash = arguments["/aes128"];
+                encType = Interop.KERB_ETYPE.aes128_cts_hmac_sha1;
+            }
+            else if (arguments.ContainsKey("/aes256"))
             {
                 hash = arguments["/aes256"];
                 encType = Interop.KERB_ETYPE.aes256_cts_hmac_sha1;
             }
+
             if (arguments.ContainsKey("/ptt"))
             {
                 ptt = true;
@@ -120,13 +141,13 @@ namespace Rubeus.Commands
             }
             if (String.IsNullOrEmpty(hash))
             {
-                Console.WriteLine("\r\n[X] You must supply a /password or /rc4 hash or /aes256 hash!\r\n");
+                Console.WriteLine("\r\n[X] You must supply a /password , or a [/des|/rc4|/aes128|/aes256] hash!\r\n");
                 return;
             }
 
-            if (!((encType == Interop.KERB_ETYPE.rc4_hmac) || (encType == Interop.KERB_ETYPE.aes256_cts_hmac_sha1)))
+            if (!((encType == Interop.KERB_ETYPE.des_cbc_md5) || (encType == Interop.KERB_ETYPE.rc4_hmac) || (encType == Interop.KERB_ETYPE.aes128_cts_hmac_sha1) || (encType == Interop.KERB_ETYPE.aes256_cts_hmac_sha1)))
             {
-                Console.WriteLine("\r\n[X] Only /rc4 and /aes256 are supported at this time.\r\n");
+                Console.WriteLine("\r\n[X] Only /des, /rc4, /aes128, and /aes256 are supported at this time.\r\n");
                 return;
             }
             else