+++ Rubeus/Domain/CommandCollection.cs
@@ -10,10 +10,10 @@ namespace Rubeus.Domain
 
         // How To Add A New Command:
         //  1. Create your command class in the Commands Folder
-        //  2. That class must have a CommandName static property that has the Command's name
-        //      and must also Implement the ICommand interface
-        //  3. Put the code that does the work into the Execute() method
-        //  4. Add an entry to the _availableCommands dictionary in the Constructor below.
+        //      a. That class must have a CommandName static property that has the Command's name
+        //              and must also Implement the ICommand interface
+        //      b. Put the code that does the work into the Execute() method
+        //  2. Add an entry to the _availableCommands dictionary in the Constructor below.
 
         public CommandCollection()
         {
@@ -24,6 +24,7 @@ namespace Rubeus.Domain
             _availableCommands.Add(Createnetonly.CommandName, () => new Createnetonly());
             _availableCommands.Add(Describe.CommandName, () => new Describe());
             _availableCommands.Add(Dump.CommandName, () => new Dump());
+            _availableCommands.Add(Hash.CommandName, () => new Hash());
             _availableCommands.Add(HarvestCommand.CommandName, () => new HarvestCommand());
             _availableCommands.Add(Kerberoast.CommandName, () => new Kerberoast());
             _availableCommands.Add(Klist.CommandName, () => new Klist());