+++ Rubeus/lib/Crypto.cs
@@ -8,79 +8,73 @@ namespace Rubeus
 {
     public class Crypto
     {
-        // Adapted from Kevin-Robertson powershell Get-KerberosAESKey: https://gist.github.com/Kevin-Robertson/9e0f8bfdbf4c1e694e6ff4197f0a4372
-        // References:
-        // * [MS-KILE] open spec' https://msdn.microsoft.com/library/cc233855.aspx?f=255&MSPPError=-2147217396
-        // * RFC regarding AES in Kerberos https://www.rfc-editor.org/rfc/pdfrfc/rfc3962.txt.pdf
-        public static byte[] ComputeAES256KerberosKey(byte[] password, byte[] salt)
+        public static void ComputeAllKerberosPasswordHashes(string password, string userName = "", string domainName = "")
         {
-            byte[] AES256_CONSTANT = { 0x6B, 0x65, 0x72, 0x62, 0x65, 0x72, 0x6F, 0x73, 0x7B, 0x9B, 0x5B, 0x2B, 0x93, 0x13, 0x2B, 0x93, 0x5C, 0x9B, 0xDC, 0xDA, 0xD9, 0x5C, 0x98, 0x99, 0xC4, 0xCA, 0xE4, 0xDE, 0xE6, 0xD6, 0xCA, 0xE4 };
-            const int rounds = 4096;
-            var pbkdf2 = new System.Security.Cryptography.Rfc2898DeriveBytes(password, salt, rounds);
-            byte[] pbkdf2_aes256_key = pbkdf2.GetBytes(32);
+            // use KerberosPasswordHash() to calculate rc4_hmac, aes128_cts_hmac_sha1, aes256_cts_hmac_sha1, and des_cbc_md5 hashes for a given password
 
-            string pbkdf2_aes256_key_string = System.BitConverter.ToString(pbkdf2_aes256_key).Replace("-", "");
+            Console.WriteLine("\r\n[*] Action: Calculate Password Hashes\r\n");
 
-            var aes = new System.Security.Cryptography.AesManaged();
-            aes.Mode = System.Security.Cryptography.CipherMode.CBC;
-            aes.Padding = System.Security.Cryptography.PaddingMode.None;
-            aes.KeySize = 256;
-            aes.Key = pbkdf2_aes256_key;
-            aes.IV = new byte[] { 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00 };
-            var encryptor = aes.CreateEncryptor();
-            byte[] aes256_key_part_1 = encryptor.TransformFinalBlock(AES256_CONSTANT, 0, AES256_CONSTANT.Length);
-            byte[] aes256_key_part_2 = encryptor.TransformFinalBlock(aes256_key_part_1, 0, aes256_key_part_1.Length);
-            byte[] aes256_key = aes256_key_part_1.Take(16).Concat(aes256_key_part_2.Take(16)).ToArray();
-            return aes256_key;
-        }
+            Console.WriteLine("[*] Input password             : {0}", password);
 
-        public static byte[] ComputeRC4KerberosKey(string password)
-        {
-            // RC4 Kerberos Key is NTLM
-            byte[] password_byte = System.Text.Encoding.Unicode.GetBytes(password);
-            return ComputeMD4(password_byte);
-        }
+            string salt = String.Format("{0}{1}", domainName.ToUpper(), userName.ToLower());
 
-        // source: https://rosettacode.org/wiki/MD4#C.23
-        private static byte[] ComputeMD4(byte[] plain)
-        {
-            // get padded uints from bytes
-            List<byte> bytes = plain.ToList();
-            uint bitCount = (uint)(bytes.Count) * 8;
-            bytes.Add(128);
-            while (bytes.Count % 64 != 56) bytes.Add(0);
-            var uints = new List<uint>();
-            for (int i = 0; i + 3 < bytes.Count; i += 4)
-                uints.Add(bytes[i] | (uint)bytes[i + 1] << 8 | (uint)bytes[i + 2] << 16 | (uint)bytes[i + 3] << 24);
-            uints.Add(bitCount);
-            uints.Add(0);
-
-            // run rounds
-            uint a = 0x67452301, b = 0xefcdab89, c = 0x98badcfe, d = 0x10325476;
-            Func<uint, uint, uint> rol = (x, y) => x << (int)y | x >> 32 - (int)y;
-            for (int q = 0; q + 15 < uints.Count; q += 16)
+            if (!String.IsNullOrEmpty(userName) && !String.IsNullOrEmpty(domainName))
             {
-                var chunk = uints.GetRange(q, 16);
-                uint aa = a, bb = b, cc = c, dd = d;
-                Action<Func<uint, uint, uint, uint>, uint[]> round = (f, y) =>
-                {
-                    foreach (uint i in new[] { y[0], y[1], y[2], y[3] })
-                    {
-                        a = rol(a + f(b, c, d) + chunk[(int)(i + y[4])] + y[12], y[8]);
-                        d = rol(d + f(a, b, c) + chunk[(int)(i + y[5])] + y[12], y[9]);
-                        c = rol(c + f(d, a, b) + chunk[(int)(i + y[6])] + y[12], y[10]);
-                        b = rol(b + f(c, d, a) + chunk[(int)(i + y[7])] + y[12], y[11]);
-                    }
-                };
-                round((x, y, z) => (x & y) | (~x & z), new uint[] { 0, 4, 8, 12, 0, 1, 2, 3, 3, 7, 11, 19, 0 });
-                round((x, y, z) => (x & y) | (x & z) | (y & z), new uint[] { 0, 1, 2, 3, 0, 4, 8, 12, 3, 5, 9, 13, 0x5a827999 });
-                round((x, y, z) => x ^ y ^ z, new uint[] { 0, 2, 1, 3, 0, 8, 4, 12, 3, 9, 11, 15, 0x6ed9eba1 });
-                a += aa; b += bb; c += cc; d += dd;
+                Console.WriteLine("[*] Input username             : {0}", userName);
+                Console.WriteLine("[*] Input domain               : {0}", domainName);
+                Console.WriteLine("[*] Salt                       : {0}", salt);
             }
 
-            // return hex encoded string
-            byte[] outBytes = new[] { a, b, c, d }.SelectMany(BitConverter.GetBytes).ToArray();
-            return outBytes;
+            string rc4Hash = KerberosPasswordHash(Interop.KERB_ETYPE.rc4_hmac, password);
+            Console.WriteLine("[*]       rc4_hmac             : {0}", rc4Hash);
+
+            if (String.IsNullOrEmpty(userName) || String.IsNullOrEmpty(domainName))
+            {
+                Console.WriteLine("\r\n[!] /user:X and /domain:Y need to be supplied to calculate AES and DES hash types!");
+            }
+            else
+            {
+                string aes128Hash = KerberosPasswordHash(Interop.KERB_ETYPE.aes128_cts_hmac_sha1, password, salt);
+                Console.WriteLine("[*]       aes128_cts_hmac_sha1 : {0}", aes128Hash);
+
+                string aes256Hash = KerberosPasswordHash(Interop.KERB_ETYPE.aes256_cts_hmac_sha1, password, salt);
+                Console.WriteLine("[*]       aes256_cts_hmac_sha1 : {0}", aes256Hash);
+
+                string desHash = KerberosPasswordHash(Interop.KERB_ETYPE.des_cbc_md5, String.Format("{0}{1}", password, salt), salt);
+                Console.WriteLine("[*]       des_cbc_md5          : {0}", desHash);
+            }
+
+            Console.WriteLine();
+        }
+
+        public static string KerberosPasswordHash(Interop.KERB_ETYPE etype, string password, string salt = "", int count = 4096)
+        {
+            // use the internal KERB_ECRYPT HashPassword() function to calculate a password hash of a given etype
+            // adapted from @gentilkiwi's Mimikatz "kerberos::hash" implementation
+
+            Interop.KERB_ECRYPT pCSystem;
+            IntPtr pCSystemPtr;
+
+            // locate the crypto system for the hash type we want
+            int status = Interop.CDLocateCSystem(etype, out pCSystemPtr);
+
+            pCSystem = (Interop.KERB_ECRYPT)System.Runtime.InteropServices.Marshal.PtrToStructure(pCSystemPtr, typeof(Interop.KERB_ECRYPT));
+            if (status != 0)
+                throw new System.ComponentModel.Win32Exception(status, "Error on CDLocateCSystem");
+
+            // get the delegate for the password hash function
+            Interop.KERB_ECRYPT_HashPassword pCSystemHashPassword = (Interop.KERB_ECRYPT_HashPassword)System.Runtime.InteropServices.Marshal.GetDelegateForFunctionPointer(pCSystem.HashPassword, typeof(Interop.KERB_ECRYPT_HashPassword));
+            Interop.UNICODE_STRING passwordUnicode = new Interop.UNICODE_STRING(password);
+            Interop.UNICODE_STRING saltUnicode = new Interop.UNICODE_STRING(salt);
+
+            byte[] output = new byte[pCSystem.KeySize];
+
+            int success = pCSystemHashPassword(passwordUnicode, saltUnicode, count, output);
+
+            if (status != 0)
+                throw new Win32Exception(status);
+
+            return System.BitConverter.ToString(output).Replace("-", "");
         }
 
         // Adapted from Vincent LE TOUX' "MakeMeEnterpriseAdmin"